+++ src/NeatMapper/Interfaces/Core/Mapper/IMergeMapFactory.cs
@@ -39,9 +39,11 @@ namespace NeatMapper {
 	/// <typeparam name="TDestination">Destination type.</typeparam>
 	/// <remarks>Implementations of this class must be thread-safe.</remarks>
 	public abstract class MergeMapFactory<TSource, TDestination> : IMergeMapFactory {
-		public abstract Type SourceType { get; }
+		// DEV: virtual to be backwards compatible, should remove
+		public virtual Type SourceType => typeof(TSource);
 
-		public abstract Type DestinationType { get; }
+		// DEV: virtual to be backwards compatible, should remove
+		public virtual Type DestinationType => typeof(TDestination);
 
 
 		/// <inheritdoc cref="IMergeMapFactory.Invoke(object?, object?)" path="/summary"/>
@@ -62,6 +64,9 @@ namespace NeatMapper {
 
 
 		object? IMergeMapFactory.Invoke(object? source, object? destination) {
+			TypeUtils.CheckObjectType(source, typeof(TSource), nameof(source));
+			TypeUtils.CheckObjectType(destination, typeof(TDestination), nameof(destination));
+
 			return Invoke((TSource?)source, (TDestination?)destination);
 		}