+++ src/NeatMapper/Interfaces/Core/Matcher/IMatchMapFactory.cs
@@ -36,9 +36,11 @@ namespace NeatMapper {
 	/// <typeparam name="TDestination">Destination type.</typeparam>
 	/// <remarks>Implementations of this class must be thread-safe.</remarks>
 	public abstract class MatchMapFactory<TSource, TDestination> : IMatchMapFactory {
-		public abstract Type SourceType { get; }
+		// DEV: virtual to be backwards compatible, should remove
+		public virtual Type SourceType => typeof(TSource);
 
-		public abstract Type DestinationType { get; }
+		// DEV: virtual to be backwards compatible, should remove
+		public virtual Type DestinationType => typeof(TDestination);
 
 
 		/// <inheritdoc cref="IMatchMapFactory.Invoke(object?, object?)"/>
@@ -53,6 +55,9 @@ namespace NeatMapper {
 
 
 		bool IMatchMapFactory.Invoke(object? source, object? destination) {
+			TypeUtils.CheckObjectType(source, typeof(TSource), nameof(source));
+			TypeUtils.CheckObjectType(destination, typeof(TDestination), nameof(destination));
+
 			return Invoke((TSource?)source, (TDestination?)destination);
 		}